refact: extract NavDrawer into stateless widgets
