+++ lib/ui/styled_components/styled_components.dart
@@ -1,3 +1,2 @@
 export 'app_logo.dart';
-export 'styled_nav_drawer.dart';
 export 'styled_protocols_year.dart';
