+++ lib/ui/views/home/home_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:wvems_protocols/ui/styled_components/styled_components.dart';
 import 'package:wvems_protocols/ui/views/home/fab/home_fab.dart';
 import 'package:wvems_protocols/ui/views/home/header/home_screen_header.dart';
 import 'package:wvems_protocols/ui/views/home/home_state_testing.dart';
+import 'package:wvems_protocols/ui/views/nav_drawer/nav_drawer.dart';
 
 import 'home_controller.dart';
 
@@ -16,7 +16,7 @@ class HomeScreen extends StatelessWidget {
       builder: (_) {
         return Scaffold(
           resizeToAvoidBottomInset: false,
-          drawer: StyledNavDrawer(),
+          drawer: NavDrawer(),
           key: _.homeScaffoldKey,
           body: SafeArea(
             child: Column(
