+++ lib/ui/views/nav_drawer/items/items.dart
@@ -1,0 +1,6 @@
+export 'about_item.dart';
+export 'download_item.dart';
+export 'new_messages_item.dart';
+export 'settings_item.dart';
+export 'share_item.dart';
+export 'version_item.dart';
