+++ lib/ui/views/nav_drawer/shared/nav_icon.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class NavIcon extends StatelessWidget {
+  const NavIcon(this.icon, {Key? key}) : super(key: key);
+
+  final IconData icon;
+
+  @override
+  Widget build(BuildContext context) {
+    return Icon(icon, size: 30.0);
+  }
+}
